+++ plugins/embedbase/dialogs/embedbase.js
@@ -29,14 +29,15 @@ CKEDITOR.dialog.add( 'embedBase', function( editor ) {
 				// Indicate visually that waiting for the response (https://dev.ckeditor.com/ticket/13213).
 				that.setState( CKEDITOR.DIALOG_STATE_BUSY );
 
-				var url = that.getValueOf( 'info', 'url' );
+				var url = that.getValueOf( 'info', 'url' ),
+					widget = that.getModel( editor );
 
-				loadContentRequest = that.widget.loadContent( url, {
+				loadContentRequest = widget.loadContent( url, {
 					noNotifications: true,
 
 					callback: function() {
-						if ( !that.widget.isReady() ) {
-							editor.widgets.finalizeCreation( that.widget.wrapper.getParent( true ) );
+						if ( !widget.isReady() ) {
+							editor.widgets.finalizeCreation( widget.wrapper.getParent( true ) );
 						}
 
 						editor.fire( 'saveSnapshot' );
@@ -48,7 +49,7 @@ CKEDITOR.dialog.add( 'embedBase', function( editor ) {
 					errorCallback: function( messageTypeOrMessage ) {
 						that.getContentElement( 'info', 'url' ).select();
 
-						alert( that.widget.getErrorMessage( messageTypeOrMessage, url, 'Given' ) );
+						alert( widget.getErrorMessage( messageTypeOrMessage, url, 'Given' ) );
 
 						unlock();
 					}
@@ -85,7 +86,9 @@ CKEDITOR.dialog.add( 'embedBase', function( editor ) {
 						},
 
 						validate: function() {
-							if ( !this.getDialog().widget.isUrlValid( this.getValue() ) ) {
+							var widget = this.getDialog().getModel( editor );
+
+							if ( !widget.isUrlValid( this.getValue() ) ) {
 								return lang.unsupportedUrlGiven;
 							}