+++ plugins/forms/dialogs/hiddenfield.js
@@ -9,18 +9,23 @@ CKEDITOR.dialog.add( 'hiddenfield', function( editor ) {
 		hiddenField: null,
 		minWidth: 350,
 		minHeight: 110,
-		onShow: function() {
-			delete this.hiddenField;
-
-			var editor = this.getParentEditor(),
-				selection = editor.getSelection(),
+		getModel: function( editor ) {
+			var selection = editor.getSelection(),
 				element = selection.getSelectedElement();
 
 			if ( element && element.data( 'cke-real-element-type' ) && element.data( 'cke-real-element-type' ) == 'hiddenfield' ) {
-				this.hiddenField = element;
-				element = editor.restoreRealElement( this.hiddenField );
-				this.setupContent( element );
-				selection.selectElement( this.hiddenField );
+				return element;
+			}
+
+			return null;
+		},
+		onShow: function() {
+			var editor = this.getParentEditor(),
+				element = this.getModel( editor );
+
+			if ( element ) {
+				this.setupContent( editor.restoreRealElement( element ) );
+				editor.getSelection().selectElement( element );
 			}
 		},
 		onOk: function() {
@@ -32,11 +37,13 @@ CKEDITOR.dialog.add( 'hiddenfield', function( editor ) {
 
 			element.setAttribute( 'type', 'hidden' );
 			this.commitContent( element );
-			var fakeElement = editor.createFakeElement( element, 'cke_hidden', 'hiddenfield' );
-			if ( !this.hiddenField )
+			var fakeElement = editor.createFakeElement( element, 'cke_hidden', 'hiddenfield' ),
+				hiddenField = this.getModel( editor );
+
+			if ( !hiddenField )
 				editor.insertElement( fakeElement );
 			else {
-				fakeElement.replace( this.hiddenField );
+				fakeElement.replace( hiddenField );
 				editor.getSelection().selectElement( fakeElement );
 			}
 			return true;