+++ plugins/image2/dialogs/image2.js
@@ -389,7 +389,7 @@ CKEDITOR.dialog.add( 'image2', function( editor ) {
 		},
 		onShow: function() {
 			// Create a "global" reference to edited widget.
-			widget = this.widget;
+			widget = this.getModel();
 
 			// Create a "global" reference to widget's image.
 			image = widget.parts.image;